Update shell.c
Fixes a few issues in the previous version and adds runCommand struct functionality. 
Generates warning on compile: comparison of array 'command.pipeIn' not equal to a null pointer is always true. We should find another way to show there is no piping. I'll do more tomorrow, but this is all for tonight